Add new contracts to toml file

+++ infrastructure/zk/src/env.ts
@@ -3,12 +3,14 @@ import fs from 'fs';
 import dotenv from 'dotenv';
 import * as utils from './utils';
 import * as config from './config';
+import * as toml from "@iarna/toml";
 
 export function get() {
     fs.readdirSync('etc/env').forEach((file) => {
-        if (file.endsWith('.bak') || file.endsWith('example') || file == 'current') {
+        if (!file.endsWith(".env")){
             return;
         }
+
         const env = file.replace(/\..*$/, '');
         if (env == process.env.ZKSYNC_ENV) {
             console.log(' * ' + env);
@@ -29,11 +31,18 @@ export async function gitHooks() {
 
 export function set(env: string) {
     const envFile = `etc/env/${env}.env`;
+    const envDir = `etc/env/${env}`;
     if (!fs.existsSync(envFile)) {
         throw new Error(envFile + ' not found');
     }
+    if (!fs.existsSync(envDir)) {
+        throw new Error(envFile + ' not found');
+    }
+
+
     fs.writeFileSync('etc/env/current', env);
     process.env.ENV_FILE = envFile;
+    process.env.ENV_DIR = envDir
     process.env.ZKSYNC_ENV = env;
     get();
 }
@@ -54,14 +63,23 @@ export async function load() {
     const zksyncEnv =
         process.env.ZKSYNC_ENV || (fs.existsSync(current) ? fs.readFileSync(current).toString().trim() : 'dev');
     const envFile = `etc/env/${zksyncEnv}.env`;
-    if (zksyncEnv == 'dev' && !fs.existsSync('etc/env/dev.env')) {
-        await config.compileConfig();
+    const envDir = `etc/env/${zksyncEnv}`;
+   if (zksyncEnv == 'dev') {
+       /// If there no folder with toml files we should delete the old dev.env and regenerate toml files and
+       if (!fs.existsSync('etc/env/dev')) {
+           fs.rmSync('etc/env/dev.env')
+       }
+
+       if (!fs.existsSync('etc/env/dev.env')) {
+           await config.compileConfig();
+       }
     }
     if (!fs.existsSync(envFile)) {
         throw new Error('ZkSync config file not found: ' + envFile);
     }
     process.env.ZKSYNC_ENV = zksyncEnv;
     process.env.ENV_FILE = envFile;
+    process.env.ENV_DIR = envDir
     dotenv.config({ path: envFile });
 }
 
@@ -72,6 +90,17 @@ export function modify(variable: string, assignedVariable: string) {
     const envFile = process.env.ENV_FILE as string;
     utils.replaceInFile(envFile, `${variable}=.*`, assignedVariable.trim());
     reload();
+}
+
+export function modify_contracts_toml(variable: string, assignedVariable: string){
+    const toml_file =`${process.env.ENV_DIR}/contracts.toml`;
+    const source = fs.readFileSync(toml_file).toString();
+    const toml_res = toml.parse(source);
+    const trimmed_variable = variable.replace("CONTRACTS_", "");
+    const trimmed_value = assignedVariable.split("=");
+    // @ts-ignore
+    toml_res["contracts"][trimmed_variable] = trimmed_value[1];
+    fs.writeFileSync(toml_file, toml.stringify(toml_res));
 }
 
 export const command = new Command('env')
